% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_type}
\alias{add_type}
\alias{add_line}
\alias{add_col}
\title{Add Plot Element}
\usage{
add_type(plot, type = NULL, ...)

add_line(plot, ...)

add_col(plot, ...)
}
\arguments{
\item{plot}{a \code{ggplot2} plot}

\item{type}{a \code{ggplot2} geom name, all geoms are supported. Full function names can be used (e.g., \code{"geom_line"}), but they can also be abbreviated (e.g., \code{"l"}, \code{"line"}). These geoms can be abbreviated by their first character: area (\code{"a"}), boxplot (\code{"b"}), column (\code{"c"}), histogram (\code{"h"}), jitter (\code{"j"}), line (\code{"l"}), point (\code{"p"}), ribbon (\code{"r"}), violin (\code{"v"}).}

\item{...}{arguments passed on to the geom function, set using \code{type}}
}
\description{
Quickly add new 'geom' to an existing \code{plot2}/\code{ggplot} model.
}
