% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admitted_patients.R
\docType{data}
\name{admitted_patients}
\alias{admitted_patients}
\title{Example Data Set with Admitted Patients}
\format{
A \link{data.frame} with 250 observations and 6 variables:
\itemize{
\item \code{date}\cr date of hospital admission
\item \code{gender}\cr gender of the patient
\item \code{age}\cr age of the patient
\item \code{age_group}\cr age group of the age of the patient, generated with \code{\link[AMR:age_groups]{AMR::age_groups()}}
\item \code{hospital}\cr ID of the hospital, from A to D
\item \code{ward}\cr type of ward, either ICU or Non-ICU
}
}
\usage{
admitted_patients
}
\description{
An auto-generated, fictitious data set containing patients admitted to hospitals.
}
\keyword{datasets}
