% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certepackage.R
\docType{package}
\name{certeplot2-package}
\alias{certeplot2}
\alias{certeplot2-package}
\title{certeplot2: A Certe R Package for Convenient Plotting}
\description{
A Certe R Package for fast and convenient plotting, by providing wrappers around 'tidyverse' packages such as 'ggplot2', while also providing plotting in the Certe organisational style. This package is part of the 'certedata' universe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://certe-medical-epidemiology.github.io/certeplot2}
  \item \url{https://github.com/certe-medical-epidemiology/certeplot2}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.berends@certe.nl}

Authors:
\itemize{
  \item Erwin E. A. Hassing \email{e.hassing@certe.nl}
}

Other contributors:
\itemize{
  \item Certe Medical Diagnostics & Advice Foundation [copyright holder, funder]
}

}
\keyword{internal}
