% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_title.R
\name{get_plot_title}
\alias{get_plot_title}
\title{Get Plot Title}
\usage{
get_plot_title(plot, valid_filename = TRUE, default = NULL)
}
\arguments{
\item{plot}{a \code{ggplot2} plot}

\item{valid_filename}{a \link{logical} to indicate whether the returned value should be a valid filename, defaults to \code{TRUE}}

\item{default}{the default value, if a plot title is absent}
}
\description{
Get the title of the plot, or a default value. If the title is not set in a plot, this function tries to generate one from the plot mapping.
}
\examples{
# plot2() uses get_plot_title() for the main title if the title is not set manually:
iris |>
  plot2()
admitted_patients |>
  plot2(age_group, n_distinct(patient_id), ward, gender)

without_title <- plot2(mtcars)
with_title <- plot2(mtcars, title = "Plotting **mpg** vs. **cyl**!")

# default is a guess:
get_plot_title(without_title)
get_plot_title(without_title, valid_filename = FALSE)
get_plot_title(with_title)
get_plot_title(with_title, valid_filename = FALSE)

# unless 'default' is set (only affects plots without title):
get_plot_title(without_title, default = "title")
get_plot_title(with_title, default = "title")
}
