% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_to_expression.R
\name{md_to_expression}
\alias{md_to_expression}
\title{Convert Markdown to Plotmath Expression}
\usage{
md_to_expression(x)
}
\arguments{
\item{x}{text to convert, only the first value will be evaluated}
}
\value{
An \link{expression} if \code{x} is length 1, or a \link{list} of expressions otherwise
}
\description{
This function converts common markdown language to an \R \link[grDevices:plotmath]{plotmath} expression. \code{\link[=plot2]{plot2()}} uses this function internally to convert plot titles and axis titles.
}
\details{
This function only supports common markdown (italic, bold, bold-italic, subscript, superscript), but also supports some additional functionalities for more advanced expressions using \R \link[grDevices:plotmath]{plotmath}. Please see \emph{Examples}.

In \code{\link[=plot2]{plot2()}}, this function can be also set to argument \code{category.labels} to print the data values as expressions:
\itemize{
\item \code{plot2(..., category.labels = md_to_expression)}
}
}
\examples{
# use '*' for italics, not '_', to prevent conflicts with variable naming
md_to_expression("this is *italic* text, this is _not italic_ text")

md_to_expression("this is **bold** text")

md_to_expression("this is ***bold and italic*** text")

# subscript and superscript can be done in HTML or markdown with curly brackets:
md_to_expression("this is some<sub>subscripted text</sub>, this is also_{subscripted} text")
md_to_expression("this is some<sup>superscripted text</sup>, this is also^{superscripted} text")

# use $...$ to use any plotmath expression as-is (see ?plotmath):
md_to_expression("text $omega$ text, $a[x]$")

mtcars |>
  plot2(mpg, hp,
        title = "*These are* the **Greek** lower $omega$ and upper $Omega$",
        x.title = "x_{mpg}",
        y.title = "y_{hp}")
        
mtcars |> 
  plot2(mpg, hp,
        title = "$f[X](x)==frac(1, sigma*sqrt(2*pi))*plain(e)^{frac(-(x-mu)^2, 2*sigma^2)}$",
        subtitle = "Some insane $widehat(plotmath)$ title")
}
