% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_layer.R
\name{move_layer}
\alias{move_layer}
\title{Move a \code{ggplot} Layer}
\usage{
move_layer(plot, move = -1, layer = length(plot$layers))
}
\arguments{
\item{plot}{a \code{ggplot} object}

\item{move}{number of layer to move \code{layer} up or down}

\item{layer}{the layer to affect, defaults to top layer}
}
\description{
Use this function to move a certain plot layer up or down. This function returns a \code{ggplot} object.
}
