% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netherlands}
\alias{netherlands}
\title{Example Geography Data Set: the Netherlands}
\format{
An \link{sf}/\link{data.frame} with 12 observations and 3 variables:
\itemize{
\item \code{province}\cr name of the Dutch province
\item \code{area_km2}\cr area in square kilometres
\item \code{geometry}\cr geometry of the province, of class sfc_MULTIPOLYGON/sfc
}
}
\usage{
netherlands
}
\description{
A data set containing the geometies of the twelve provinces of the Netherlands, according to Statistics Netherlands (2021).
}
\keyword{datasets}
