% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elements.R
\name{plot_elements}
\alias{plot_elements}
\alias{get_theme}
\alias{get_mapping}
\alias{get_data}
\alias{get_layers}
\title{Get Plot Element}
\usage{
get_theme(x)

get_mapping(x)

get_data(x)

get_layers(x)
}
\arguments{
\item{x}{a \code{ggplot} model}
}
\description{
These functions extract \code{ggplot} properties.
}
\examples{
x <- plot2(iris)

get_mapping(x)

class(get_theme(x))
}
