% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{as_plotly}
\alias{as_plotly}
\alias{plotly_style}
\title{Create Interactive Plotly}
\usage{
as_plotly(plot, ...)

plotly_style(plot, ...)
}
\arguments{
\item{plot}{a \code{ggplot2} plot}

\item{...}{In case of \code{\link[=as_plotly]{as_plotly()}}: arguments to pass on to \code{\link[plotly:layout]{layout()}} to change the Plotly layout object

In case of \code{\link[=plotly_style]{plotly_style()}}: arguments to pass on to \code{\link[plotly:style]{style()}} to change the Plotly style object}
}
\description{
Transform a \code{ggplot2}/\code{plot2} object to an interactive plot using the \href{https://plotly.com/r/}{Plotly R Open Source Graphing Library}.
}
\examples{
mtcars \%>\%
  plot2(mpg, hp) \%>\% 
  as_plotly()
  
mtcars \%>\%
  plot2(mpg, hp) \%>\% 
  as_plotly(dragmode = "pan") \%>\%
  plotly_style(marker.line.color = "red",
               hoverinfo = "y")
}
