% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_minimal2.R
\name{theme_minimal2}
\alias{theme_minimal2}
\title{An Even More Minimal Theme}
\usage{
theme_minimal2(...)
}
\arguments{
\item{...}{arguments passed on to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\description{
This \code{ggplot2} theme provide even more white area and less clutter than \code{\link[ggplot2:ggtheme]{theme_minimal()}}.
}
\examples{
plot2(iris)
plot2(admitted_patients, x = hospital, category = gender)

if (require("ggplot2")) {
  ggplot2(mtcars, aes(hp, mpg)) +
    geom_point()
}
if (require("ggplot2")) {
  ggplot2(mtcars, aes(hp, mpg)) +
    geom_point() +
    theme_minimal2()
}
}
