% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3.R
\name{uses_decimal_comma}
\alias{uses_decimal_comma}
\title{Check if \emph{locale} uses comma as decimal separator}
\source{
\url{https://en.wikipedia.org/wiki/Decimal_mark#Countries_using_Arabic_numerals_with_decimal_comma}
}
\usage{
uses_decimal_comma()
}
\value{
logical
}
\description{
This function returns \code{TRUE} if \code{\link[=Sys.getlocale]{Sys.getlocale()}} contains one of the countries where a comma "," is used as decimal separator (see Source). This function returns \code{FALSE} otherwise.
}
