% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_type.R
\name{add_type}
\alias{add_type}
\alias{add_line}
\alias{add_col}
\title{Add Plot Element}
\usage{
add_type(plot, type = NULL, mapping = aes(), ...)

add_line(
  plot,
  y,
  x,
  group = 1,
  colour = "certeblauw",
  size,
  linetype,
  ...,
  inherit.aes = TRUE
)

add_col(
  plot,
  y,
  x,
  colour = "certeblauw",
  colour_fill = "certeblauw",
  size,
  width,
  ...,
  inherit.aes = TRUE
)
}
\arguments{
\item{plot}{a \code{ggplot2} plot}

\item{type}{a \code{ggplot2} geom name, all geoms are supported. Full function names can be used (e.g., \code{"geom_line"}), but they can also be abbreviated (e.g., \code{"l"}, \code{"line"}). These geoms can be abbreviated by their first character: area (\code{"a"}), boxplot (\code{"b"}), column (\code{"c"}), histogram (\code{"h"}), jitter (\code{"j"}), line (\code{"l"}), point (\code{"p"}), ribbon (\code{"r"}), violin (\code{"v"}).}

\item{mapping}{a mapping created with \code{\link[ggplot2:aes]{ggplot2::aes()}} to pass on to the geom}

\item{x, y}{aesthetic arguments}

\item{group, linetype, size, width, ...}{arguments passed on to the geom}

\item{colour, colour_fill}{colour of the line or column, will be evaluated with \code{\link[certestyle:colourpicker]{certestyle::colourpicker()}}. If \code{colour_fill} is missing but \code{colour} is given, \code{colour_fill} will inherit the colour set with \code{colour}.}

\item{inherit.aes}{a \link{logical} to indicate whether the default aesthetics should be inherited, rather than combining with them}
}
\description{
Quickly add a new 'geom' to an existing \code{plot2}/\code{ggplot} model. Like \code{\link[=plot2]{plot2()}}, they support tidy evaluation, meaning that variables can be unquoted. They can be added using the pipe (\verb{\%>\%}).
}
\examples{
df <- data.frame(var_1 = c(1:100),
                 var_2 = rnorm(100, 100, 25))
df \%>\%
  plot2() \%>\% 
  add_line(mean(var_2))
  
df \%>\%
  plot2() \%>\% 
  add_line(y = mean(var_2), 
           size = 2) \%>\%
  add_col(y = var_2 / 5,
          width = 0.25,
          colour = "certeroze")
   
if (require("certestats", warn.conflicts = FALSE)) {
   df \%>\%
     plot2(caption = "EWMA in pink :)") \%>\% 
     add_line(y = ewma(var_2, 0.75),
              colour = "certeroze",
              linetype = 3,
              size = 1)
}
}
