% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labellers.R
\name{labellers}
\alias{labellers}
\alias{label_euro}
\title{Label Euro currencies}
\usage{
label_euro(
  accuracy = NULL,
  scale = 1,
  prefix = "€",
  suffix = "",
  big.mark = big_mark(),
  decimal.mark = dec_mark(),
  trim = TRUE,
  largest_with_cents = 1e+05,
  ...
)
}
\arguments{
\item{accuracy, largest_with_cents}{Number to round to. If \code{NULL}, the default,
values will be rounded to the nearest integer, unless any of the
values has non-zero fractional component (e.g. cents) and the largest
value is less than \code{largest_with_cents} which by default is 100,000.}

\item{scale}{A scaling factor: \code{x} will be multiplied by \code{scale} before
formatting. This is useful if the underlying data is very small or very
large.}

\item{prefix, suffix}{Symbols to display before and after value.}

\item{big.mark}{Character used between every 3 digits to separate thousands.}

\item{decimal.mark}{The character to be used to indicate the numeric
decimal point.}

\item{trim}{Logical, if \code{FALSE}, values are right-justified to a common
width (see \code{\link[base:format]{base::format()}}).}

\item{...}{Other arguments passed on to \code{\link[base:format]{base::format()}}.}
}
\description{
Format numbers as currency, rounding values to dollars or cents using a convenient heuristic. This function is similar to \code{\link[=label_dollar]{label_dollar()}} from the \code{scales} package.
}
