% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_to_expression.R
\name{md_to_expression}
\alias{md_to_expression}
\title{Convert Markdown to Plotmath Expression}
\usage{
md_to_expression(x)
}
\arguments{
\item{x}{text to convert, only the first value will be evaluated}
}
\value{
An \link{expression}
}
\description{
This function converts common markdown language to an \R \link[grDevices:plotmath]{plotmath} expression. \code{\link[=plot2]{plot2()}} uses this function internally to convert plot titles and axis titles.
}
\details{
This function only supports common markdown (italic, bold, bold-italic, subscript, superscript), but also supports some additional functionalities for more advanced expressions using \R \link[grDevices:plotmath]{plotmath}. Please see \emph{Examples}.

\emph{Rationale}: we tried to use the \code{ggtext} package instead to support markdown using their \code{element_markdown()} function for \code{ggplot2} \link[ggplot2:theme]{themes}, but it appeared to be incompatible with the \code{showtext} package to support different fonts. We subsequently developed this function to transform markdown into \link[grDevices:plotmath]{plotmath}.
}
\examples{
# use '*' for italics, not '_', to prevent conflicts with variable naming
md_to_expression("this is *italic* text, this is _not italic_ text")

md_to_expression("this is **bold** text")

md_to_expression("this is ***bold and italic*** text")

# subscript and superscript can never be preceded by a space.
md_to_expression("this is long<sub>subscripted text</sub>, this is short_subscripted text")
md_to_expression("this is long<sup>superscripted text</sup>, this is short^superscripted text")

# use $...$ to use any plotmath expression as-is (see ?plotmath):
md_to_expression("text $omega$ text, $a[x]$")

mtcars |>
  plot2(mpg, hp,
        title = "These are the **Greek** lower $omega$ and upper $Omega$")
        
mtcars |> 
  plot2(mpg, hp,
        title = "$f[X](x)==frac(1, sigma*sqrt(2*pi))*plain(e)^{frac(-(x-mu)^2, 2*sigma^2)}$",
        subtitle = "Some insane $widehat(plotmath)$ title")
}
