% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{plot2-methods}
\alias{plot2-methods}
\alias{plot2.default}
\alias{plot2.numeric}
\alias{plot2.freq}
\alias{plot2.sf}
\alias{plot2.data.frame}
\title{Methods for \code{\link[=plot2]{plot2()}}}
\usage{
\method{plot2}{default}(.data, ...)

\method{plot2}{numeric}(.data, ...)

\method{plot2}{freq}(.data, x = .data$item, y = .data$count, x.sort = "freq-desc", ...)

\method{plot2}{sf}(
  .data,
  x = NULL,
  y = NULL,
  colour = "grey25",
  colour_fill = c("white", "certeblauw"),
  x.expand = 0,
  y.expand = 0,
  x.title = NULL,
  y.title = NULL,
  datalabels = FALSE,
  datalabels.colour = "black",
  size = 0.1,
  legend.position = "right",
  legend.title = TRUE,
  legend.reverse = TRUE,
  theme = theme_minimal2(panel.grid.major = element_blank(), panel.grid.minor =
    element_blank(), panel.border = element_blank(), axis.title = element_blank(),
    axis.text = element_blank(), axis.line = element_blank(), axis.ticks =
    element_blank()),
  ...
)

\method{plot2}{data.frame}(
  .data = NULL,
  x = NULL,
  y = NULL,
  category = NULL,
  facet = NULL,
  geom = NULL,
  x.title = NULL,
  y.title = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  title.linelength = 60,
  title.colour = "black",
  subtitle.linelength = 60,
  subtitle.colour = "grey35",
  na.replace = "(??)",
  na.rm = FALSE,
  facet.position = "top",
  facet.fill = NULL,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  facet.repeat_lbls_x = TRUE,
  facet.repeat_lbls_y = FALSE,
  facet.fixed_y = FALSE,
  facet.drop = FALSE,
  facet.nrow = NULL,
  facet.relative = FALSE,
  x.date_breaks = NULL,
  x.date_labels = NULL,
  category.focus = NULL,
  colour = "certe",
  colour_fill = NULL,
  colour.opacity = 0,
  x.lbl_angle = 0,
  x.lbl_align = NULL,
  x.lbl_italic = FALSE,
  x.remove = FALSE,
  x.position = "bottom",
  x.max_items = Inf,
  x.max_txt = "(rest, x \%n)",
  category.max_items = Inf,
  category.max_txt = "(rest, x \%n)",
  facet.max_items = Inf,
  facet.max_txt = "(rest, x \%n)",
  x.breaks = NULL,
  x.breaks_n = NULL,
  x.trans = "identity",
  x.expand = 0.5,
  x.limits = NULL,
  x.character = NULL,
  y.remove = FALSE,
  y.24h = FALSE,
  y.age = FALSE,
  y.percent = FALSE,
  y.percent_break = 10,
  y.breaks = NULL,
  y.limits = NULL,
  y.labels = NULL,
  y.expand = 0.25,
  y.trans = "identity",
  y.position = "left",
  category.labels = NULL,
  category.percent = FALSE,
  category.breaks = NULL,
  category.limits = NULL,
  category.expand = 0,
  category.trans = "identity",
  x.sort = NULL,
  category.sort = TRUE,
  facet.sort = TRUE,
  datalabels = TRUE,
  datalabels.round = ifelse(y.percent, 2, 1),
  datalabels.colour = "grey25",
  datalabels.fill = NULL,
  datalabels.size = (3 * text_factor),
  datalabels.angle = 0,
  decimal.mark = ",",
  big.mark = ".",
  summarise_function = base::sum,
  stacked = FALSE,
  stackedpercent = FALSE,
  horizontal = FALSE,
  reverse = horizontal,
  smooth = FALSE,
  smooth.method = NULL,
  smooth.formula = NULL,
  smooth.se = TRUE,
  smooth.level = 0.95,
  smooth.alpha = 0.15,
  smooth.size = 0.75,
  smooth.linetype = 3,
  size = NULL,
  linetype = 1,
  bins = NULL,
  width = NULL,
  jitter_seed = NA,
  violin_scale = "count",
  legend.position = "top",
  legend.title = FALSE,
  legend.reverse = FALSE,
  legend.barheight = 5,
  legend.barwidth = 1,
  legend.nbin = 300,
  legend.italic = FALSE,
  zoom = FALSE,
  sep = "/",
  print = FALSE,
  text_factor = 1,
  family = "Calibri",
  theme = theme_minimal2(horizontal = horizontal),
  markdown = TRUE,
  x.category = NULL,
  y.category = NULL,
  ...
)
}
\arguments{
\item{.data}{data to plot}
}
\description{
These are the implemented methods for different S3 classes to be used in \code{\link[=plot2]{plot2()}}.
}
