% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plot2}
\alias{get_plot_title}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{plot2}{\code{\link[plot2]{get_plot_title}}, \code{\link[plot2]{plot2}}}
}}

