% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labellers.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{label_dollar}
\alias{n}
\alias{n_distinct}
\alias{everything}
\alias{starts_with}
\alias{ends_with}
\alias{matches}
\alias{where}
\alias{first}
\alias{last}
\alias{all_of}
\alias{any_of}
\alias{dec_mark}
\alias{big_mark}
\alias{colourpicker}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{certestyle}{\code{\link[certestyle:dec_mark]{big_mark}}, \code{\link[certestyle]{colourpicker}}, \code{\link[certestyle]{dec_mark}}}

  \item{dplyr}{\code{\link[dplyr:reexports]{all_of}}, \code{\link[dplyr:reexports]{any_of}}, \code{\link[dplyr:nth]{first}}, \code{\link[dplyr:nth]{last}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr]{n_distinct}}}

  \item{scales}{\code{\link[scales]{label_dollar}}}

  \item{tidyselect}{\code{\link[tidyselect:starts_with]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect:starts_with]{matches}}, \code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{where}}}
}}

