% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_minimal2.R
\name{theme_minimal2}
\alias{theme_minimal2}
\title{An Even More Minimal Theme}
\usage{
theme_minimal2(...)
}
\arguments{
\item{...}{arguments passed on to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\description{
An Even More Minimal Theme
}
\examples{
plot2(iris)
plot2(iris, x = Species)
}
