% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_minimal2}
\alias{theme_minimal2}
\title{An Even More Minimal Theme}
\usage{
theme_minimal2(
  subtitle.colour = "black",
  x.lbl.angle = 0,
  x.lbl.align = 0.5,
  horizontal = FALSE,
  legend.position = "top",
  legend.italic = FALSE,
  font.size_factor = 1,
  facet.fill = NA,
  facet.bold = TRUE,
  facet.italic = FALSE,
  facet.size = 10,
  facet.margin = 8,
  has_subtitle = FALSE,
  ...
)
}
\description{
An Even More Minimal Theme
}
\examples{
plot2(iris)
plot2(iris, x = Species)
}
